add tabs to image builder table
